refactor!: move DelegateStateMixin to component-base
