+++ packages/accordion/src/vaadin-accordion-panel.d.ts
@@ -3,9 +3,9 @@
  * Copyright (c) 2019 - 2022 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
+import { DelegateStateMixin } from '@vaadin/component-base/src/delegate-state-mixin.js';
 import { DetailsMixin } from '@vaadin/details/src/vaadin-details-mixin.js';
 import { DelegateFocusMixin } from '@vaadin/field-base/src/delegate-focus-mixin.js';
-import { DelegateStateMixin } from '@vaadin/field-base/src/delegate-state-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
 
 /**
