+++ packages/component-base/src/delegate-state-mixin.d.ts
