+++ packages/tabsheet/package.json
@@ -37,7 +37,6 @@
   "dependencies": {
     "@polymer/polymer": "^3.0.0",
     "@vaadin/component-base": "24.0.0-alpha7",
-    "@vaadin/field-base": "24.0.0-alpha7",
     "@vaadin/scroller": "24.0.0-alpha7",
     "@vaadin/tabs": "24.0.0-alpha7",
     "@vaadin/vaadin-lumo-styles": "24.0.0-alpha7",
